cleaned up code and fixed authenticator
Took 3 hours 48 minutes
