+++ app/src/main/java/com/dsd/kosjenka/di/AdapterModule.kt
@@ -13,96 +13,5 @@ import com.dsd.kosjenka.databinding.UserProfileListItemBinding
 import com.dsd.kosjenka.domain.models.UserProfile
 
 object AdapterModule {
-    class UserProfilesAdapter(
-        private val mListener: ProfileItemClickListener
-    ) : RecyclerView.Adapter<UserProfilesAdapter.UserProfileViewHolder>(){
 
-        class UserProfileViewHolder(
-            val binding: UserProfileListItemBinding
-        ) : RecyclerView.ViewHolder(binding.root) {
-            fun bind(currentProfile: UserProfile, listener: ProfileItemClickListener, isSpecialItem: Boolean){
-                if (isSpecialItem) {
-                    binding.profileListName.text = "Add Profile"
-                    binding.profileImageView.setImageResource(R.drawable.add_profile_btn_view)
-                    binding.editProfileButton.visibility = ImageView.GONE
-                    binding.root.setOnClickListener {
-                        listener.let {
-                            it.onAddProfileClick()
-                        }
-                    }
-                    binding.executePendingBindings()
-                }
-                else {
-                    binding.profileListName.text = currentProfile.username
-                    binding.profileImageView.setImageResource(R.drawable.user_avatar_1)
-                    binding.root.setOnClickListener {
-                        listener.onProfileClick(
-                            currentProfile
-                        )
-                    }
-//                    binding.root.setOnLongClickListener{
-//                        listener.onLongProfileClick(currentProfile)
-//                        true
-//                    }
-                    binding.editProfileButton.setOnClickListener{
-                        listener.onEditProfileClick(currentProfile)
-                    }
-                    binding.executePendingBindings()
-                }
-            }
-
-            companion object {
-                fun from(parent: ViewGroup): UserProfileViewHolder {
-                    val layoutInflater = LayoutInflater.from(parent.context)
-                    val itemBinding =
-                        UserProfileListItemBinding
-                            .inflate(layoutInflater, parent, false)
-                    return UserProfileViewHolder(itemBinding)
-                }
-            }
-        }
-
-        override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): UserProfileViewHolder {
-            return UserProfileViewHolder.from(parent)
-        }
-
-        override fun getItemCount(): Int {
-            return differ.currentList.size + 1
-        }
-
-        override fun onBindViewHolder(holder: UserProfileViewHolder, position: Int) {
-            val isSpecialItem : Boolean = (position == differ.currentList.size)
-
-            if (isSpecialItem){
-                holder.bind(UserProfile(-1, 0,"", 0.0), mListener, isSpecialItem)
-            } else {
-                holder.bind(differ.currentList[position], mListener, isSpecialItem)
-            }
-
-        }
-
-        private val differCallback = object : DiffUtil.ItemCallback<UserProfile>() {
-            override fun areItemsTheSame(
-                oldItem: UserProfile,
-                newItem: UserProfile,
-            ): Boolean =
-                oldItem.id_user == newItem.id_user && oldItem.username == newItem.username
-
-            override fun areContentsTheSame(
-                oldItem: UserProfile,
-                newItem: UserProfile,
-            ): Boolean =
-                oldItem.id_account == newItem.id_account
-                        && oldItem.proficiency == newItem.proficiency
-        }
-
-        val differ = AsyncListDiffer(this, differCallback)
-
-        interface ProfileItemClickListener{
-            fun onProfileClick(profile: UserProfile)
-            fun onAddProfileClick()
-            fun onEditProfileClick(profile: UserProfile)
-            fun onLongProfileClick(profile: UserProfile)
-        }
-    }
 }
