+++ app/src/main/res/values/strings.xml
@@ -34,4 +34,6 @@
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="empty_list">There are no results for your search!</string>
+    <string name="token_error">Token expired! Please sign in again.</string>
+    <string name="add_profile_btn">Add Profile</string>
 </resources>
